# Improve `PandasIO.insert()` performance