Applied to pandas.Series with a single name (by far the most common use case) and more than 100 elements. This is about ~11 times faster in a synthetic bench translating unique 20 IDs, randomly distributed in the series (table below).

| Size        | µ ± σ (old)       | µ ± σ (new)       |
|-------------|-------------------|-------------------|
| 10          | 186 µs ± 4.42 µs  | 228 µs ± 2.66 µs  |
| 1'000       | 354 µs ± 4.85 µs  | 259 µs ± 4.8 µs   |
| 10'000      | 1.86 ms ± 3.15 µs | 353 µs ± 4 µs     |
| 100'000     | 18.3 ms ± 125 µs  | 1.36 ms ± 16.5 µs |
| 500'000     | 97.5 ms ± 637 µs  | 8.49 ms ± 260 µs  |
| 1'000'000   | 190 ms ± 564 µs   | 16.7 ms ± 206 µs  |
| 5'000'000   | 948 ms ± 4.26 ms  | 90.4 ms ± 656 µs  |
| 25'000'000  | 4.8 s ± 29.8 ms   | 433 ms ± 6.84 ms  |
| 250'000'000 | 47.2 s ± 358 ms   | 4.18 s ± 28.9 ms  |